# Conflicts:
#	src/main/java/GDG/whatssue/domain/club/controller/ClubController.java
#	src/main/java/GDG/whatssue/domain/club/entity/Club.java
#	src/main/java/GDG/whatssue/domain/schedule/entity/Schedule.java
#	src/main/java/GDG/whatssue/domain/user/entity/User.java
#	src/main/java/GDG/whatssue/global/argumentresolver/LoginUserArgumentResolver.java

+++ src/main/java/GDG/whatssue/domain/file/service/impl/S3UploadService.java
@@ -1,6 +1,11 @@
 package GDG.whatssue.domain.file.service.impl;
 
+import GDG.whatssue.domain.file.entity.ClubProfileImage;
+import GDG.whatssue.domain.file.entity.MemberProfileImage;
+import GDG.whatssue.domain.file.entity.PostImage;
+import GDG.whatssue.domain.file.entity.UploadFile;
 import GDG.whatssue.domain.file.service.FileUploadService;
+import GDG.whatssue.domain.file.FileConst;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -10,6 +15,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 @Service
 @RequiredArgsConstructor
@@ -23,21 +29,25 @@ public class S3UploadService implements FileUploadService {
     @Value("${cloud.aws.s3.bucket}")
     public String bucket;
 
-
     @Override
-    public String saveFile(MultipartFile multipartFile, String dirName) throws IOException {
-        String originalFileName = multipartFile.getOriginalFilename();
+    public String uploadFile(MultipartFile multipartFile, String dirName) throws IOException {
+        String storeFileName;
 
-        String fileName = getFileName(dirName, originalFileName);
+        //저장 사진이 없으면 기본 경로 반환
+        if (multipartFile == null || multipartFile.isEmpty()) {
+            return storeFileName = dirName + FileConst.DEFAULT_IMAGE_NAME;
+        }
+
+        storeFileName = getFileName(dirName, multipartFile.getOriginalFilename());
 
         ObjectMetadata metadata = new ObjectMetadata();
         metadata.setContentLength(multipartFile.getSize());
         metadata.setContentType(multipartFile.getContentType());
 
-        amazonS3.putObject(new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), metadata)
+        amazonS3.putObject(new PutObjectRequest(bucket, storeFileName, multipartFile.getInputStream(), metadata)
             .withCannedAcl(CannedAccessControlList.PublicRead));
 
-        return fileName;
+        return storeFileName;
     }
 
     @Override
@@ -48,12 +58,25 @@ public class S3UploadService implements FileUploadService {
 
     @Override
     public void deleteFile(String storeFileName) {
-        //TODO
+        //default 파일이면 버킷에서 삭제하지 않는다
+        if (StringUtils.substringMatch(storeFileName, 0, "default")) {
+            return;
+        }
+
         amazonS3.deleteObject(bucket, storeFileName);
     }
 
-    public String getFullPath(String fileName) {
-        return PATH + fileName;
+    @Override
+    public String getOriginalFileName(MultipartFile multipartFile) {
+        String originalFileName;
+
+        if (multipartFile == null || multipartFile.isEmpty()) {
+            originalFileName = "default";
+        } else {
+            originalFileName = multipartFile.getOriginalFilename();
+        }
+
+        return originalFileName;
     }
 
     private String getFileName(String dirName, String originalFileName) {
